svrctl(2) update
- synchronize request type with ioctl by making it unsigned long;
- unbreak VFS requests, as they were being sent to PM;
- use proper ioctl direction flags (and new numbers) for requests;
- remove some needless header inclusions;
- svrctl is in libc, make its message name reflect this;
- keep backward compatibility: svrctl is part of the userland ABI.

+++ minix/commands/svrctl/svrctl.c
@@ -20,7 +20,7 @@ static char *bin_name;
 
 int main (int argc, char *argv[])
 {
-  int param;
+  unsigned long param;
   endpoint_t proc_e = NONE;
   struct sysgetenv sysgetenv;
   char *to_whom, *operation, *what, *value;
